[travis] Support building with the locally installed CMake on test host

+++ scripts/install-deps.sh
@@ -43,7 +43,7 @@ function main() {
 
     mkdir -p "$prefix"
 
-    if [ ! -f "$prefix/bin/cmake" ]; then
+    if [ ! -f "$prefix/bin/cmake" -a -z "${USE_LOCAL_CMAKE}" ]; then
         if [ ! -f "cmake-3.10.0-rc5-Linux-x86_64.sh" ]; then
             curl -L -O "https://cmake.org/files/v3.10/cmake-3.10.0-rc5-Linux-x86_64.sh"
         fi
